server/requirements.txt
selenium==4.5.0
webdriver_manager==3.8.4
airium==0.2.5
packaging==21.3
PyYAML==6.0
googlemaps==4.6.0
Flask==2.2.5
Pillow==9.3.0
protobuf==4.21.6
requests==2.28.1
Werkzeug==2.2.3
paho-mqtt==1.6.1
numpy==1.24.3